#57 Initial refactor, moving all items into their respective controller, view or module package Program cannot locate "Dashboard.fxml" when starting up the program
