+++ javadoc/allclasses-index.html
@@ -101,7 +101,7 @@ loadScripts(document, 'script');</script>
 </th>
 </tr>
 <tr class="row-color" id="i5">
-<td class="col-first"><a href="com/simulator/gui/Logger.html" title="class in com.simulator.gui">Logger</a></td>
+<td class="col-first"><a href="com/simulator/gui/Logger.html" title="class in com.simulator.view">Logger</a></td>
 <th class="col-last" scope="row">
 <div class="block">This is a singleton class used to hold the outputConsole so it can be written to by all that need it</div>
 </th>
@@ -143,7 +143,7 @@ loadScripts(document, 'script');</script>
 </th>
 </tr>
 <tr class="alt-color" id="i12">
-<td class="col-first"><a href="com/simulator/gui/RoomControlsController.html" title="class in com.simulator.gui">RoomControlsController</a></td>
+<td class="col-first"><a href="com/simulator/gui/RoomControlsController.html" title="class in com.simulator.view">RoomControlsController</a></td>
 <th class="col-last" scope="row">
 <div class="block">This is the controller class for the RoomControls.fxml file</div>
 </th>
@@ -161,7 +161,7 @@ loadScripts(document, 'script');</script>
 </th>
 </tr>
 <tr class="row-color" id="i15">
-<td class="col-first"><a href="com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.gui">SimulationParameterController</a></td>
+<td class="col-first"><a href="com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.view">SimulationParameterController</a></td>
 <th class="col-last" scope="row">
 <div class="block">This is the controller class for the ParametorEditor.fxml file</div>
 </th>
@@ -173,13 +173,13 @@ loadScripts(document, 'script');</script>
 </th>
 </tr>
 <tr class="row-color" id="i17">
-<td class="col-first"><a href="com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.gui">SmartHomeSimulator</a></td>
+<td class="col-first"><a href="com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.view">SmartHomeSimulator</a></td>
 <th class="col-last" scope="row">
 <div class="block">This is the launch class of the simulator.</div>
 </th>
 </tr>
 <tr class="alt-color" id="i18">
-<td class="col-first"><a href="com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.gui">SmartHomeSimulatorController</a></td>
+<td class="col-first"><a href="com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.view">SmartHomeSimulatorController</a></td>
 <th class="col-last" scope="row">
 <div class="block">This is the controller class for the RoomControls.fxml file AND the Dashboard.fxml file</div>
 </th>
