+++ javadoc/allpackages-index.html
@@ -64,7 +64,7 @@ loadScripts(document, 'script');</script>
 </thead>
 <tbody>
 <tr class="alt-color">
-<th class="col-first" scope="row"><a href="com/simulator/gui/package-summary.html">com.simulator.gui</a></th>
+<th class="col-first" scope="row"><a href="com/simulator/gui/package-summary.html">com.simulator.view</a></th>
 <td class="col-last">&nbsp;</td>
 </tr>
 <tr class="row-color">
