+++ javadoc/com/simulator/gui/Logger.html
@@ -6,7 +6,7 @@
 <title>Logger</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2020-11-12">
-<meta name="description" content="declaration: package: com.simulator.gui, class: Logger">
+<meta name="description" content="declaration: package: com.simulator.view, class: Logger">
 <meta name="generator" content="javadoc/ClassWriterImpl">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../script-dir/jquery-ui.min.css" title="Style">
@@ -73,11 +73,11 @@ loadScripts(document, 'script');</script>
 <main role="main">
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">com.simulator.gui</a></div>
+<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">com.simulator.view</a></div>
 <h1 title="Class Logger" class="title">Class Logger</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree">java.lang.Object
-<div class="inheritance">com.simulator.gui.Logger</div>
+<div class="inheritance">com.simulator.controller.Logger</div>
 </div>
 <section class="description">
 <hr>
@@ -104,14 +104,14 @@ extends java.lang.Object</pre>
 </thead>
 <tbody>
 <tr class="alt-color" id="i0">
-<td class="col-first"><code>static <a href="Logger.html" title="class in com.simulator.gui">Logger</a></code></td>
+<td class="col-first"><code>static <a href="Logger.html" title="class in com.simulator.view">Logger</a></code></td>
 <th class="col-second" scope="row"><code><span class="member-name-link"><a href="#getInstance()">getInstance</a></span>()</code></th>
 <td class="col-last">
 <div class="block">returns the instance that is held</div>
 </td>
 </tr>
 <tr class="row-color" id="i1">
-<td class="col-first"><code>static <a href="Logger.html" title="class in com.simulator.gui">Logger</a></code></td>
+<td class="col-first"><code>static <a href="Logger.html" title="class in com.simulator.view">Logger</a></code></td>
 <th class="col-second" scope="row"><code><span class="member-name-link"><a href="#newInstance(javafx.scene.control.TextArea)">newInstance</a></span>&#8203;(javafx.scene.control.TextArea&nbsp;outputConsole2)</code></th>
 <td class="col-last">
 <div class="block">creates a new instance of the Logger object to be contained within the instance Logger object</div>
@@ -159,7 +159,7 @@ extends java.lang.Object</pre>
 <li>
 <section class="detail" id="newInstance(javafx.scene.control.TextArea)">
 <h3>newInstance</h3>
-<div class="member-signature"><span class="modifiers">public static</span>&nbsp;<span class="return-type"><a href="Logger.html" title="class in com.simulator.gui">Logger</a></span>&nbsp;<span class="member-name">newInstance</span>&#8203;(<span class="parameters">javafx.scene.control.TextArea&nbsp;outputConsole2)</span></div>
+<div class="member-signature"><span class="modifiers">public static</span>&nbsp;<span class="return-type"><a href="Logger.html" title="class in com.simulator.view">Logger</a></span>&nbsp;<span class="member-name">newInstance</span>&#8203;(<span class="parameters">javafx.scene.control.TextArea&nbsp;outputConsole2)</span></div>
 <div class="block">creates a new instance of the Logger object to be contained within the instance Logger object</div>
 <dl class="notes">
 <dt>Parameters:</dt>
@@ -172,7 +172,7 @@ extends java.lang.Object</pre>
 <li>
 <section class="detail" id="getInstance()">
 <h3>getInstance</h3>
-<div class="member-signature"><span class="modifiers">public static</span>&nbsp;<span class="return-type"><a href="Logger.html" title="class in com.simulator.gui">Logger</a></span>&nbsp;<span class="member-name">getInstance</span>()</div>
+<div class="member-signature"><span class="modifiers">public static</span>&nbsp;<span class="return-type"><a href="Logger.html" title="class in com.simulator.view">Logger</a></span>&nbsp;<span class="member-name">getInstance</span>()</div>
 <div class="block">returns the instance that is held</div>
 <dl class="notes">
 <dt>Returns:</dt>
