+++ javadoc/com/simulator/gui/SmartHomeSimulatorController.html
@@ -6,7 +6,7 @@
 <title>SmartHomeSimulatorController</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2020-11-12">
-<meta name="description" content="declaration: package: com.simulator.gui, class: SmartHomeSimulatorController">
+<meta name="description" content="declaration: package: com.simulator.view, class: SmartHomeSimulatorController">
 <meta name="generator" content="javadoc/ClassWriterImpl">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../script-dir/jquery-ui.min.css" title="Style">
@@ -73,11 +73,11 @@ loadScripts(document, 'script');</script>
 <main role="main">
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">com.simulator.gui</a></div>
+<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">com.simulator.view</a></div>
 <h1 title="Class SmartHomeSimulatorController" class="title">Class SmartHomeSimulatorController</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree">java.lang.Object
-<div class="inheritance">com.simulator.gui.SmartHomeSimulatorController</div>
+<div class="inheritance">com.simulator.controller.SmartHomeSimulatorController</div>
 </div>
 <section class="description">
 <hr>
