+++ javadoc/com/simulator/gui/package-summary.html
@@ -3,10 +3,10 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (15) on Thu Nov 12 12:15:22 EST 2020 -->
-<title>com.simulator.gui</title>
+<title>com.simulator.view</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2020-11-12">
-<meta name="description" content="declaration: package: com.simulator.gui">
+<meta name="description" content="declaration: package: com.simulator.view">
 <meta name="generator" content="javadoc/PackageWriterImpl">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../script-dir/jquery-ui.min.css" title="Style">
@@ -51,7 +51,7 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;com.simulator.gui</h1>
+<h1 title="Package" class="title">Package&nbsp;com.simulator.view</h1>
 </div>
 <section class="summary">
 <ul class="summary-list">
@@ -67,31 +67,31 @@ loadScripts(document, 'script');</script>
 </thead>
 <tbody>
 <tr class="alt-color">
-<th class="col-first" scope="row"><a href="Logger.html" title="class in com.simulator.gui">Logger</a></th>
+<th class="col-first" scope="row"><a href="Logger.html" title="class in com.simulator.view">Logger</a></th>
 <td class="col-last">
 <div class="block">This is a singleton class used to hold the outputConsole so it can be written to by all that need it</div>
 </td>
 </tr>
 <tr class="row-color">
-<th class="col-first" scope="row"><a href="RoomControlsController.html" title="class in com.simulator.gui">RoomControlsController</a></th>
+<th class="col-first" scope="row"><a href="RoomControlsController.html" title="class in com.simulator.view">RoomControlsController</a></th>
 <td class="col-last">
 <div class="block">This is the controller class for the RoomControls.fxml file</div>
 </td>
 </tr>
 <tr class="alt-color">
-<th class="col-first" scope="row"><a href="SimulationParameterController.html" title="class in com.simulator.gui">SimulationParameterController</a></th>
+<th class="col-first" scope="row"><a href="SimulationParameterController.html" title="class in com.simulator.view">SimulationParameterController</a></th>
 <td class="col-last">
 <div class="block">This is the controller class for the ParametorEditor.fxml file</div>
 </td>
 </tr>
 <tr class="row-color">
-<th class="col-first" scope="row"><a href="SmartHomeSimulator.html" title="class in com.simulator.gui">SmartHomeSimulator</a></th>
+<th class="col-first" scope="row"><a href="SmartHomeSimulator.html" title="class in com.simulator.view">SmartHomeSimulator</a></th>
 <td class="col-last">
 <div class="block">This is the launch class of the simulator.</div>
 </td>
 </tr>
 <tr class="alt-color">
-<th class="col-first" scope="row"><a href="SmartHomeSimulatorController.html" title="class in com.simulator.gui">SmartHomeSimulatorController</a></th>
+<th class="col-first" scope="row"><a href="SmartHomeSimulatorController.html" title="class in com.simulator.view">SmartHomeSimulatorController</a></th>
 <td class="col-last">
 <div class="block">This is the controller class for the RoomControls.fxml file AND the Dashboard.fxml file</div>
 </td>
