+++ javadoc/com/simulator/gui/package-tree.html
@@ -3,10 +3,10 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (15) on Thu Nov 12 12:15:22 EST 2020 -->
-<title>com.simulator.gui Class Hierarchy</title>
+<title>com.simulator.view Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2020-11-12">
-<meta name="description" content="tree: package: com.simulator.gui">
+<meta name="description" content="tree: package: com.simulator.view">
 <meta name="generator" content="javadoc/PackageTreeWriter">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../script-dir/jquery-ui.min.css" title="Style">
@@ -51,7 +51,7 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 class="title">Hierarchy For Package com.simulator.gui</h1>
+<h1 class="title">Hierarchy For Package com.simulator.view</h1>
 <span class="package-hierarchy-label">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../../overview-tree.html">All Packages</a></li>
@@ -64,13 +64,13 @@ loadScripts(document, 'script');</script>
 <ul>
 <li class="circle">javafx.application.Application
 <ul>
-<li class="circle">com.simulator.gui.<a href="SmartHomeSimulator.html" title="class in com.simulator.gui"><span class="type-name-link">SmartHomeSimulator</span></a></li>
+<li class="circle">com.simulator.view.<a href="SmartHomeSimulator.html" title="class in com.simulator.view"><span class="type-name-link">SmartHomeSimulator</span></a></li>
 </ul>
 </li>
-<li class="circle">com.simulator.gui.<a href="Logger.html" title="class in com.simulator.gui"><span class="type-name-link">Logger</span></a></li>
-<li class="circle">com.simulator.gui.<a href="RoomControlsController.html" title="class in com.simulator.gui"><span class="type-name-link">RoomControlsController</span></a> (implements javafx.fxml.Initializable)</li>
-<li class="circle">com.simulator.gui.<a href="SimulationParameterController.html" title="class in com.simulator.gui"><span class="type-name-link">SimulationParameterController</span></a></li>
-<li class="circle">com.simulator.gui.<a href="SmartHomeSimulatorController.html" title="class in com.simulator.gui"><span class="type-name-link">SmartHomeSimulatorController</span></a></li>
+<li class="circle">com.simulator.view.<a href="Logger.html" title="class in com.simulator.view"><span class="type-name-link">Logger</span></a></li>
+<li class="circle">com.simulator.view.<a href="RoomControlsController.html" title="class in com.simulator.view"><span class="type-name-link">RoomControlsController</span></a> (implements javafx.fxml.Initializable)</li>
+<li class="circle">com.simulator.view.<a href="SimulationParameterController.html" title="class in com.simulator.view"><span class="type-name-link">SimulationParameterController</span></a></li>
+<li class="circle">com.simulator.view.<a href="SmartHomeSimulatorController.html" title="class in com.simulator.view"><span class="type-name-link">SmartHomeSimulatorController</span></a></li>
 </ul>
 </li>
 </ul>
