+++ javadoc/element-list
@@ -1,2 +1,2 @@
-com.simulator.gui
+com.simulator.view
 com.simulator.model
