+++ javadoc/index-files/index-10.html
@@ -56,7 +56,7 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">N</a>&nbsp;<a href="index-11.html">O</a>&nbsp;<a href="index-12.html">P</a>&nbsp;<a href="index-13.html">R</a>&nbsp;<a href="index-14.html">S</a>&nbsp;<a href="index-15.html">T</a>&nbsp;<a href="index-16.html">U</a>&nbsp;<a href="index-17.html">V</a>&nbsp;<a href="index-18.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:N">N</h2>
 <dl class="index">
-<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#newInstance(javafx.scene.control.TextArea)">newInstance(TextArea)</a></span> - Static method in class com.simulator.gui.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.gui">Logger</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#newInstance(javafx.scene.control.TextArea)">newInstance(TextArea)</a></span> - Static method in class com.simulator.view.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.view">Logger</a></dt>
 <dd>
 <div class="block">creates a new instance of the Logger object to be contained within the instance Logger object</div>
 </dd>
