+++ javadoc/index-files/index-11.html
@@ -62,11 +62,11 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/simulator/model/Observer.html#%3Cinit%3E()">Observer()</a></span> - Constructor for class com.simulator.model.<a href="../com/simulator/model/Observer.html" title="class in com.simulator.model">Observer</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#ouputToConsole(java.lang.String)">ouputToConsole(String)</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.gui">Logger</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#ouputToConsole(java.lang.String)">ouputToConsole(String)</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.view">Logger</a></dt>
 <dd>
 <div class="block">outputs a given string to the Output Console on the Smart home dashboard</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#outputToLogFile(java.lang.String)">outputToLogFile(String)</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.gui">Logger</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#outputToLogFile(java.lang.String)">outputToLogFile(String)</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.view">Logger</a></dt>
 <dd>
 <div class="block">Outputs to log.txt file</div>
 </dd>
