+++ javadoc/index-files/index-13.html
@@ -60,7 +60,7 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">Remove a light</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#resetLogFile()">resetLogFile()</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.gui">Logger</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#resetLogFile()">resetLogFile()</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.view">Logger</a></dt>
 <dd>
 <div class="block">resets the contents of the log.txt file</div>
 </dd>
@@ -70,11 +70,11 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/simulator/model/Room.html#%3Cinit%3E(java.lang.String,java.lang.String)">Room(String, String)</a></span> - Constructor for class com.simulator.model.<a href="../com/simulator/model/Room.html" title="class in com.simulator.model">Room</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/simulator/gui/RoomControlsController.html" title="class in com.simulator.gui"><span class="type-name-link">RoomControlsController</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.gui</a></dt>
+<dt><a href="../com/simulator/gui/RoomControlsController.html" title="class in com.simulator.view"><span class="type-name-link">RoomControlsController</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.view</a></dt>
 <dd>
 <div class="block">This is the controller class for the RoomControls.fxml file</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/RoomControlsController.html#%3Cinit%3E()">RoomControlsController()</a></span> - Constructor for class com.simulator.gui.<a href="../com/simulator/gui/RoomControlsController.html" title="class in com.simulator.gui">RoomControlsController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/RoomControlsController.html#%3Cinit%3E()">RoomControlsController()</a></span> - Constructor for class com.simulator.view.<a href="../com/simulator/gui/RoomControlsController.html" title="class in com.simulator.view">RoomControlsController</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">N</a>&nbsp;<a href="index-11.html">O</a>&nbsp;<a href="index-12.html">P</a>&nbsp;<a href="index-13.html">R</a>&nbsp;<a href="index-14.html">S</a>&nbsp;<a href="index-15.html">T</a>&nbsp;<a href="index-16.html">U</a>&nbsp;<a href="index-17.html">V</a>&nbsp;<a href="index-18.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a></main>
