+++ javadoc/index-files/index-14.html
@@ -181,29 +181,29 @@ loadScripts(document, 'script');</script>
 <div class="block">Set the User Location
  Check if auto lights need to be turned on</div>
 </dd>
-<dt><a href="../com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.gui"><span class="type-name-link">SimulationParameterController</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.gui</a></dt>
+<dt><a href="../com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.view"><span class="type-name-link">SimulationParameterController</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.view</a></dt>
 <dd>
 <div class="block">This is the controller class for the ParametorEditor.fxml file</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SimulationParameterController.html#%3Cinit%3E()">SimulationParameterController()</a></span> - Constructor for class com.simulator.gui.<a href="../com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.gui">SimulationParameterController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SimulationParameterController.html#%3Cinit%3E()">SimulationParameterController()</a></span> - Constructor for class com.simulator.view.<a href="../com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.view">SimulationParameterController</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="../com/simulator/model/SimulationParameters.html" title="class in com.simulator.model"><span class="type-name-link">SimulationParameters</span></a> - Class in <a href="../com/simulator/model/package-summary.html">com.simulator.model</a></dt>
 <dd>
 <div class="block">A simulationParameters class using the singleton pattern to create one instance that can be accessed anywhere</div>
 </dd>
-<dt><a href="../com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.gui"><span class="type-name-link">SmartHomeSimulator</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.gui</a></dt>
+<dt><a href="../com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.view"><span class="type-name-link">SmartHomeSimulator</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.view</a></dt>
 <dd>
 <div class="block">This is the launch class of the simulator.</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulator.html#%3Cinit%3E()">SmartHomeSimulator()</a></span> - Constructor for class com.simulator.gui.<a href="../com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.gui">SmartHomeSimulator</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulator.html#%3Cinit%3E()">SmartHomeSimulator()</a></span> - Constructor for class com.simulator.view.<a href="../com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.view">SmartHomeSimulator</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.gui"><span class="type-name-link">SmartHomeSimulatorController</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.gui</a></dt>
+<dt><a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.view"><span class="type-name-link">SmartHomeSimulatorController</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.view</a></dt>
 <dd>
 <div class="block">This is the controller class for the RoomControls.fxml file AND the Dashboard.fxml file</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#%3Cinit%3E()">SmartHomeSimulatorController()</a></span> - Constructor for class com.simulator.gui.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.gui">SmartHomeSimulatorController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#%3Cinit%3E()">SmartHomeSimulatorController()</a></span> - Constructor for class com.simulator.view.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.view">SmartHomeSimulatorController</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulator.html#start(javafx.stage.Stage)">start(Stage)</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.gui">SmartHomeSimulator</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulator.html#start(javafx.stage.Stage)">start(Stage)</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.view">SmartHomeSimulator</a></dt>
 <dd>
 <div class="block">Launch the dashboard of the simulator.</div>
 </dd>
