+++ javadoc/index-files/index-2.html
@@ -56,7 +56,7 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">N</a>&nbsp;<a href="index-11.html">O</a>&nbsp;<a href="index-12.html">P</a>&nbsp;<a href="index-13.html">R</a>&nbsp;<a href="index-14.html">S</a>&nbsp;<a href="index-15.html">T</a>&nbsp;<a href="index-16.html">U</a>&nbsp;<a href="index-17.html">V</a>&nbsp;<a href="index-18.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:C">C</h2>
 <dl class="index">
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#changeAwayStatus()">changeAwayStatus()</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.gui">SmartHomeSimulatorController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#changeAwayStatus()">changeAwayStatus()</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.view">SmartHomeSimulatorController</a></dt>
 <dd>
 <div class="block">Changes the away status of the simulator.</div>
 </dd>
@@ -70,7 +70,7 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/simulator/model/USER_TYPE.html#CHILD">CHILD</a></span> - com.simulator.model.<a href="../com/simulator/model/USER_TYPE.html" title="enum in com.simulator.model">USER_TYPE</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/simulator/gui/package-summary.html">com.simulator.gui</a> - package com.simulator.gui</dt>
+<dt><a href="../com/simulator/gui/package-summary.html">com.simulator.view</a> - package com.simulator.view</dt>
 <dd>&nbsp;</dd>
 <dt><a href="../com/simulator/model/package-summary.html">com.simulator.model</a> - package com.simulator.model</dt>
 <dd>&nbsp;</dd>
