+++ javadoc/index-files/index-5.html
@@ -62,7 +62,7 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/simulator/model/SimulationParameters.html#getAllUserNames()">getAllUserNames()</a></span> - Method in class com.simulator.model.<a href="../com/simulator/model/SimulationParameters.html" title="class in com.simulator.model">SimulationParameters</a></dt>
 <dd>
-<div class="block">Get all user names in a List<String></div>
+    <div class="block">Get all user names in a List<String></String></div>
 </dd>
 <dt><span class="member-name-link"><a href="../com/simulator/model/SimulationParameters.html#getAllUsers()">getAllUsers()</a></span> - Method in class com.simulator.model.<a href="../com/simulator/model/SimulationParameters.html" title="class in com.simulator.model">SimulationParameters</a></dt>
 <dd>
@@ -110,7 +110,7 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/simulator/model/Room.html#getId()">getId()</a></span> - Method in class com.simulator.model.<a href="../com/simulator/model/Room.html" title="class in com.simulator.model">Room</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#getInstance()">getInstance()</a></span> - Static method in class com.simulator.gui.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.gui">Logger</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/Logger.html#getInstance()">getInstance()</a></span> - Static method in class com.simulator.view.<a href="../com/simulator/gui/Logger.html" title="class in com.simulator.view">Logger</a></dt>
 <dd>
 <div class="block">returns the instance that is held</div>
 </dd>
@@ -190,7 +190,7 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">return a list of all the names of the rooms</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#getSimToggle()">getSimToggle()</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.gui">SmartHomeSimulatorController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#getSimToggle()">getSimToggle()</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.view">SmartHomeSimulatorController</a></dt>
 <dd>
 <div class="block">return the toggle button</div>
 </dd>
