+++ javadoc/index-files/index-7.html
@@ -60,15 +60,15 @@ loadScripts(document, 'script');</script>
 <dd>&nbsp;</dd>
 <dt><span class="member-name-link"><a href="../com/simulator/model/PERMISSION_TYPE.html#IN_ROOM">IN_ROOM</a></span> - com.simulator.model.<a href="../com/simulator/model/PERMISSION_TYPE.html" title="enum in com.simulator.model">PERMISSION_TYPE</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SimulationParameterController.html#initialize()">initialize()</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.gui">SimulationParameterController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SimulationParameterController.html#initialize()">initialize()</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.view">SimulationParameterController</a></dt>
 <dd>
 <div class="block">Initializes the SimulationsParamterController popout window</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#initialize()">initialize()</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.gui">SmartHomeSimulatorController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/SmartHomeSimulatorController.html#initialize()">initialize()</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.view">SmartHomeSimulatorController</a></dt>
 <dd>
 <div class="block">Initializes the SmartHomeSimulator Dashboard</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/simulator/gui/RoomControlsController.html#initialize(java.net.URL,java.util.ResourceBundle)">initialize(URL, ResourceBundle)</a></span> - Method in class com.simulator.gui.<a href="../com/simulator/gui/RoomControlsController.html" title="class in com.simulator.gui">RoomControlsController</a></dt>
+<dt><span class="member-name-link"><a href="../com/simulator/gui/RoomControlsController.html#initialize(java.net.URL,java.util.ResourceBundle)">initialize(URL, ResourceBundle)</a></span> - Method in class com.simulator.view.<a href="../com/simulator/gui/RoomControlsController.html" title="class in com.simulator.view">RoomControlsController</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">N</a>&nbsp;<a href="index-11.html">O</a>&nbsp;<a href="index-12.html">P</a>&nbsp;<a href="index-13.html">R</a>&nbsp;<a href="index-14.html">S</a>&nbsp;<a href="index-15.html">T</a>&nbsp;<a href="index-16.html">U</a>&nbsp;<a href="index-17.html">V</a>&nbsp;<a href="index-18.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a></main>
