+++ javadoc/index-files/index-8.html
@@ -64,7 +64,7 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">Contructor for a new light object</div>
 </dd>
-<dt><a href="../com/simulator/gui/Logger.html" title="class in com.simulator.gui"><span class="type-name-link">Logger</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.gui</a></dt>
+<dt><a href="../com/simulator/gui/Logger.html" title="class in com.simulator.view"><span class="type-name-link">Logger</span></a> - Class in <a href="../com/simulator/gui/package-summary.html">com.simulator.view</a></dt>
 <dd>
 <div class="block">This is a singleton class used to hold the outputConsole so it can be written to by all that need it</div>
 </dd>
