+++ javadoc/overview-tree.html
@@ -54,7 +54,7 @@ loadScripts(document, 'script');</script>
 <h1 class="title">Hierarchy For All Packages</h1>
 <span class="package-hierarchy-label">Package Hierarchies:</span>
 <ul class="horizontal">
-<li><a href="com/simulator/gui/package-tree.html">com.simulator.gui</a>, </li>
+<li><a href="com/simulator/gui/package-tree.html">com.simulator.view</a>, </li>
 <li><a href="com/simulator/model/package-tree.html">com.simulator.model</a></li>
 </ul>
 </div>
@@ -65,7 +65,7 @@ loadScripts(document, 'script');</script>
 <ul>
 <li class="circle">javafx.application.Application
 <ul>
-<li class="circle">com.simulator.gui.<a href="com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.gui"><span class="type-name-link">SmartHomeSimulator</span></a></li>
+<li class="circle">com.simulator.view.<a href="com/simulator/gui/SmartHomeSimulator.html" title="class in com.simulator.view"><span class="type-name-link">SmartHomeSimulator</span></a></li>
 </ul>
 </li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Entryway.html" title="class in com.simulator.model"><span class="type-name-link">Entryway</span></a>
@@ -76,7 +76,7 @@ loadScripts(document, 'script');</script>
 </li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/House.html" title="class in com.simulator.model"><span class="type-name-link">House</span></a></li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Light.html" title="class in com.simulator.model"><span class="type-name-link">Light</span></a></li>
-<li class="circle">com.simulator.gui.<a href="com/simulator/gui/Logger.html" title="class in com.simulator.gui"><span class="type-name-link">Logger</span></a></li>
+<li class="circle">com.simulator.view.<a href="com/simulator/gui/Logger.html" title="class in com.simulator.view"><span class="type-name-link">Logger</span></a></li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Observer.html" title="class in com.simulator.model"><span class="type-name-link">Observer</span></a>
 <ul>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/SecurityModule.html" title="class in com.simulator.model"><span class="type-name-link">SecurityModule</span></a></li>
@@ -85,16 +85,16 @@ loadScripts(document, 'script');</script>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Permission.html" title="class in com.simulator.model"><span class="type-name-link">Permission</span></a></li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Profile.html" title="class in com.simulator.model"><span class="type-name-link">Profile</span></a></li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Room.html" title="class in com.simulator.model"><span class="type-name-link">Room</span></a></li>
-<li class="circle">com.simulator.gui.<a href="com/simulator/gui/RoomControlsController.html" title="class in com.simulator.gui"><span class="type-name-link">RoomControlsController</span></a> (implements javafx.fxml.Initializable)</li>
+<li class="circle">com.simulator.view.<a href="com/simulator/gui/RoomControlsController.html" title="class in com.simulator.view"><span class="type-name-link">RoomControlsController</span></a> (implements javafx.fxml.Initializable)</li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Sensor.html" title="class in com.simulator.model"><span class="type-name-link">Sensor</span></a>
 <ul>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/EntrywaySensor.html" title="class in com.simulator.model"><span class="type-name-link">EntrywaySensor</span></a></li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/MotionSensor.html" title="class in com.simulator.model"><span class="type-name-link">MotionSensor</span></a></li>
 </ul>
 </li>
-<li class="circle">com.simulator.gui.<a href="com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.gui"><span class="type-name-link">SimulationParameterController</span></a></li>
+<li class="circle">com.simulator.view.<a href="com/simulator/gui/SimulationParameterController.html" title="class in com.simulator.view"><span class="type-name-link">SimulationParameterController</span></a></li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/SimulationParameters.html" title="class in com.simulator.model"><span class="type-name-link">SimulationParameters</span></a></li>
-<li class="circle">com.simulator.gui.<a href="com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.gui"><span class="type-name-link">SmartHomeSimulatorController</span></a></li>
+<li class="circle">com.simulator.view.<a href="com/simulator/gui/SmartHomeSimulatorController.html" title="class in com.simulator.view"><span class="type-name-link">SmartHomeSimulatorController</span></a></li>
 <li class="circle">com.simulator.model.<a href="com/simulator/model/Time.html" title="class in com.simulator.model"><span class="type-name-link">Time</span></a></li>
 </ul>
 </li>
