+++ src/com/simulator/model/SecurityModule.java
@@ -1,11 +1,9 @@
 package com.simulator.model;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import com.simulator.gui.Logger;
+import com.simulator.controller.Logger;
 
 import javafx.scene.control.ToggleButton;
 /**
