+++ src/com/simulator/view/avatar_placeholder.png
