+++ src/com/simulator/view/away.png
