+++ src/com/simulator/view/closedwindow.png
