+++ src/com/simulator/view/lightOn.jpg
