+++ src/com/simulator/view/locked.png
