+++ src/com/simulator/view/openwindow.png
