+++ src/com/simulator/view/people.jpg
