+++ src/com/simulator/view/person.png
