+++ tests/com/simulator/SmartHomeSecurityModuleTest.java
@@ -1,14 +1,10 @@
 package com.simulator;
 
-import com.simulator.gui.Logger;
 import com.simulator.model.*;
 import javafx.fxml.FXML;
 import javafx.scene.control.ToggleButton;
-import javafx.scene.control.TextArea;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-
-import java.awt.*;
 
 import static org.junit.jupiter.api.Assertions.*;
